[FLINK-39669][table-planner] Extract buildWrappingProgram helper to shared CorrelateProjectionUtils
BatchPhysicalCorrelateRule and StreamPhysicalCorrelateRule each had a
byte-identical copy of the new buildWrappingProgram helper introduced in
the preceding commit. Pull it into a new object
CorrelateProjectionUtils under plan.rules.physical.common alongside the
existing CommonLookupJoinRule, and have both rules call it.

No functional change. 373 UNNEST + Correlate tests pass.
